Fixme 2: Moved PlayerInfo to the right class to make clear that it is a POJO and needed as well
