+++ src/main/java/data/PlayerInfo.java
@@ -1,4 +1,4 @@
-package model;
+package data;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
