+++ src/test/java/at/aau/serg/monopoly/websoket/GameWebSocketIntegrationTest.java
@@ -1,7 +1,7 @@
 package at.aau.serg.monopoly.websoket;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import model.PlayerInfo;
+import data.PlayerInfo;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
